requirements.txt
pandas
matplotlib
seaborn
scikit-learn
streamlit
langchain
langchain-openai
python-dotenv
langchain-experimental
tabulate  